Deleted unnecessary interfaces (#44)
