+++ src/main/java/uk/gov/ons/ctp/response/action/export/service/TemplateMappingService.java
@@ -1,58 +1,74 @@
 package uk.gov.ons.ctp.response.action.export.service;
 
+import java.util.Date;
+import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
+import java.util.stream.Collectors;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
 import uk.gov.ons.ctp.common.error.CTPException;
 import uk.gov.ons.ctp.response.action.export.domain.TemplateMapping;
+import uk.gov.ons.ctp.response.action.export.repository.TemplateMappingRepository;
 
-/** Service responsible for dealing with stored TemplateMappings. */
-public interface TemplateMappingService {
+/** The implementation of the TemplateMappingService */
+@Service
+@Slf4j
+public class TemplateMappingService {
 
-  /**
-   * To store TemplateMappings
-   *
-   * @param actionType action type of TemplateMappings
-   * @param templateMappingList the TemplateMappings json content to be stored
-   * @return the stored TemplateMappings
-   * @throws CTPException if error storing TemplateMappings
-   */
-  List<TemplateMapping> storeTemplateMappings(
-      String actionType, List<TemplateMapping> templateMappingList) throws CTPException;
+  public static final String EXCEPTION_STORE_TEMPLATE_MAPPING =
+      "Issue storing TemplateMapping. Appears to be empty.";
 
-  /**
-   * To retrieve a given TemplateMapping
-   *
-   * @param actionType for which to retrieve the TemplateMapping
-   * @return the given TemplateMapping
-   * @throws CTPException if the TemplateMapping is not found.
-   */
-  TemplateMapping retrieveTemplateMappingByActionType(String actionType) throws CTPException;
+  public static final String EXCEPTION_RETRIEVING_TEMPLATE_MAPPING = "TemplateMapping not found.";
 
-  /**
-   * To retrieve all TemplateMappings
-   *
-   * @return a list of TemplateMappings
-   */
-  List<TemplateMapping> retrieveAllTemplateMappings();
+  @Autowired private TemplateMappingRepository repository;
 
-  /**
-   * To retrieve a Map of TemplateMappings by filename.
-   *
-   * @return the Map of TemplateMappings by filename.
-   */
-  Map<String, List<TemplateMapping>> retrieveAllTemplateMappingsByFilename();
+  public List<TemplateMapping> storeTemplateMappings(
+      String actionType, List<TemplateMapping> templateMappingList) throws CTPException {
 
-  /**
-   * To retrieve a Map of TemplateMappings by actionType.
-   *
-   * @return the Map of TemplateMappings by actionType.
-   */
-  Map<String, TemplateMapping> retrieveAllTemplateMappingsByActionType();
+    for (TemplateMapping templateMapping : templateMappingList) {
+      templateMapping.setActionType(actionType);
+      templateMapping.setDateModified(new Date());
+      repository.save(templateMapping);
+    }
 
-  /**
-   * Return a list of distinct actionTypes in the collection
-   *
-   * @return a list of actionTypes.
-   */
-  List<String> retrieveActionTypes();
+    return templateMappingList;
+  }
+
+  public TemplateMapping retrieveTemplateMappingByActionType(String actionType)
+      throws CTPException {
+    TemplateMapping templateMapping = repository.findOne(actionType);
+    if (templateMapping == null) {
+      log.error("No template mapping for actionType {} found.", actionType);
+      throw new CTPException(
+          CTPException.Fault.RESOURCE_NOT_FOUND,
+          String.format("%s %s", EXCEPTION_RETRIEVING_TEMPLATE_MAPPING, actionType));
+    }
+    return templateMapping;
+  }
+
+  public List<TemplateMapping> retrieveAllTemplateMappings() {
+    return repository.findAll();
+  }
+
+  public Map<String, List<TemplateMapping>> retrieveAllTemplateMappingsByFilename() {
+    return retrieveAllTemplateMappings()
+        .stream()
+        .collect(Collectors.groupingBy(TemplateMapping::getFileNamePrefix));
+  }
+
+  public Map<String, TemplateMapping> retrieveAllTemplateMappingsByActionType() {
+    Map<String, TemplateMapping> mappings = new HashMap<String, TemplateMapping>();
+    retrieveAllTemplateMappings()
+        .forEach(
+            (templateMapping) -> {
+              mappings.put(templateMapping.getActionType(), templateMapping);
+            });
+    return mappings;
+  }
+
+  public List<String> retrieveActionTypes() {
+    return repository.findAllActionType();
+  }
 }
